Removed gordon surface from CTiglBSplineAlgorithms
Also, I refactored the pure gordon surface code
into an own class.

Closes #305
