+++ src/geometry/CTiglBSplineAlgorithms.cpp
@@ -321,23 +321,6 @@ namespace
         return array2GetRow<TColgp_Array2OfPnt, TColgp_HArray1OfPnt, Handle_TColgp_HArray1OfPnt>(matrix, rowIndex);
     }
 
-    std::vector<double> toVector(const TColStd_Array1OfReal& array)
-    {
-        std::vector<double> result(static_cast<size_t>(array.Length()));
-        int low = array.Lower();
-        for (int i = array.Lower(); i <= array.Upper(); ++i) {
-            result[static_cast<size_t>(i-low)] = array.Value(i);
-        }
-        return result;
-    }
-
-    void assertRange(const Handle(Geom_Curve)& curve, double umin, double umax, double tol=1e-7)
-    {
-        if (std::abs(curve->FirstParameter() - umin) > tol || std::abs(curve->LastParameter() - umax) > tol) {
-            throw tigl::CTiglError("Curve not in range [" + tigl::std_to_string(umin) + ", " + tigl::std_to_string(umax) + "].");
-        }
-    }
-
     void clampBSpline(Handle(Geom_BSplineCurve)& curve)
     {
         if (!curve->IsPeriodic()) {
@@ -618,170 +601,6 @@ Handle(Geom_BSplineSurface) CTiglBSplineAlgorithms::curvesToSurface(const std::v
     return CTiglBSplineAlgorithms::curvesToSurface(compatibleSplines, parameters.second, continuousIfClosed);
 }
 
-Handle(Geom_BSplineCurve) CTiglBSplineAlgorithms::reparametrizeBSpline(const Handle(Geom_BSplineCurve) spline, const TColStd_Array1OfReal& old_parameters, const TColStd_Array1OfReal& new_parameters)
-{
-    // stop, if the sizes of old_parameters and new_parameters are not equal
-    if (old_parameters.Length() != new_parameters.Length()) {
-        throw tigl::CTiglError("The amounts of old and new parameters are not equal!");
-    }
-
-    // create a copy of the given B-spline curve to avoid shadow effects
-    Handle(Geom_BSplineCurve) copied_spline = Handle(Geom_BSplineCurve)::DownCast(spline->Copy());
-
-    // sort parameter arrays
-    std::vector<double> old_parameters_vector(toVector(old_parameters));
-    std::vector<double> new_parameters_vector(toVector(new_parameters));
-
-    std::sort(old_parameters_vector.begin(), old_parameters_vector.end());
-    std::sort(new_parameters_vector.begin(), new_parameters_vector.end());
-
-    // insert knots at old parameter values, so that one can reparametrize intervals between these old parameter values afterwards
-    for (unsigned int i = 0; i < old_parameters_vector.size(); ++i) {
-        // if between parameter range of B-spline
-        double parm = old_parameters_vector[i];
-        if (parm - copied_spline->Knot(1) > 1e-15 && std::abs(copied_spline->Knot(copied_spline->NbKnots()) - parm) > 1e-15) {
-            // insert knot degree-times
-            CurveAdapterView(copied_spline).insertKnot(parm, copied_spline->Degree(), 1e-7);
-        }
-    }
-
-    // move knots to reparametrize
-    TColStd_Array1OfReal knots(1, copied_spline->NbKnots());
-    copied_spline->Knots(knots);
-    for (unsigned int interval_idx = 0; interval_idx < old_parameters_vector.size() - 1; ++interval_idx) {
-        double interval_beginning_old = old_parameters_vector[interval_idx];
-        double interval_end_old = old_parameters_vector[interval_idx + 1];
-
-        double interval_beginning_new = new_parameters_vector[interval_idx];
-        double interval_end_new = new_parameters_vector[interval_idx + 1];
-
-        for (int knot_idx = 1; knot_idx <= copied_spline->NbKnots(); ++knot_idx) {
-            double knot = copied_spline->Knot(knot_idx);
-            if (interval_beginning_old <= knot && knot < interval_end_old) {
-                knots(knot_idx) = (knot - interval_beginning_old) / (interval_end_old - interval_beginning_old) * (interval_end_new - interval_beginning_new) + interval_beginning_new;
-            }
-        }
-    }
-
-    // case: knot = upper bound of parameter range
-    knots(copied_spline->NbKnots()) = new_parameters_vector[new_parameters_vector.size() - 1];
-
-    // make sure that the new knots are unique
-    for (int knot_idx = 2; knot_idx <= knots.Length(); ++knot_idx) {
-        if (std::abs(knots(knot_idx - 1) - knots(knot_idx)) < 1e-15) {
-            throw tigl::CTiglError("There are duplicate knots in the knot vector!");
-        }
-    }
-
-    // eliminate small inaccuracy of first knot
-    if (std::abs(knots(1)) < 1e-5) {
-        knots(1) = 0.;
-    }
-
-    // the multiplicity of the knots is not modified by reparametrization itself, but by knot insertion above (degree times):
-    copied_spline->SetKnots(knots);
-
-    return copied_spline;
-}
-
-Handle(Geom_BSplineCurve) CTiglBSplineAlgorithms::reparametrizeBSplineContinuously(const Handle(Geom_BSplineCurve) spline,
-                                                                                   const TColStd_Array1OfReal& old_parameters,
-                                                                                   const TColStd_Array1OfReal& new_parameters)
-{
-    // stop, if the sizes of old_parameters and new_parameters are not equal
-    if (old_parameters.Length() != new_parameters.Length()) {
-        throw tigl::CTiglError("The amounts of old and new parameters are not equal!");
-    }
-
-    // create a copy of the given B-spline curve to avoid shadow effects
-    Handle(Geom_BSplineCurve) copied_spline = Handle(Geom_BSplineCurve)::DownCast(spline->Copy());
-
-    // sort parameter arrays
-    std::vector<double> old_parameters_vector;
-    for (int i = 1; i <= old_parameters.Length(); ++i) {
-        old_parameters_vector.push_back(old_parameters(i));
-    }
-
-    std::vector<double> new_parameters_vector;
-    for (int i = 1; i <= new_parameters.Length(); ++i) {
-        new_parameters_vector.push_back(new_parameters(i));
-    }
-
-    std::sort(old_parameters_vector.begin(), old_parameters_vector.end());
-    std::sort(new_parameters_vector.begin(), new_parameters_vector.end());
-
-    ParametrizingFunction paramFunc(old_parameters, new_parameters);
-
-    // get the flat knots
-    TColStd_Array1OfReal knots(1, copied_spline->NbKnots());
-    copied_spline->Knots(knots);
-
-    TColStd_Array1OfInteger mults(1, copied_spline->NbKnots());
-    copied_spline->Multiplicities(mults);
-
-    int n_flat_knots = 0;
-    for (int i = 1; i <= mults.Length(); ++i) {
-        n_flat_knots += mults(i);
-    }
-    TColStd_Array1OfReal flat_knots(1, n_flat_knots);
-    BSplCLib::KnotSequence(knots, mults, flat_knots);
-
-    TColgp_Array1OfPnt cp(1, copied_spline->NbPoles());
-    copied_spline->Poles(cp);
-
-    TColStd_Array1OfReal new_flat_knots(1, n_flat_knots + 5);
-    int u_degree = 3;
-    TColgp_Array1OfPnt new_cp(1, copied_spline->NbPoles() + 5);
-    int status = 0;
-
-    BSplCLib::FunctionReparameterise(paramFunc, copied_spline->Degree(), flat_knots, cp, new_flat_knots, u_degree, new_cp, status);
-
-    // find out the new knots and their multiplicities
-    int n_knots = BSplCLib::KnotsLength(new_flat_knots);
-
-    TColStd_Array1OfReal new_knots(1, n_knots);
-    TColStd_Array1OfInteger new_mults(1, n_knots);
-
-    BSplCLib::Knots(new_flat_knots, new_knots, new_mults);
-
-    // create the reparametrized B-spline
-    Handle(Geom_BSplineCurve) reparam_spline = new Geom_BSplineCurve(new_cp, new_knots, new_mults, u_degree);
-
-
-    return reparam_spline;
-}
-
-ParametrizingFunction::ParametrizingFunction(const TColStd_Array1OfReal& old_params, const TColStd_Array1OfReal& new_params)
-{
-    old_parameters = new TColStd_HArray1OfReal(1, old_params.Length());
-    new_parameters = new TColStd_HArray1OfReal(1, new_params.Length());
-    for (int param_idx = 1; param_idx <= old_params.Length(); ++param_idx) {
-        old_parameters->SetValue(param_idx, old_params(param_idx));
-        new_parameters->SetValue(param_idx, new_params(param_idx));
-    }
-}
-
-void ParametrizingFunction::Evaluate(const Standard_Integer theDerivativeRequest, const Standard_Real * theStartEnd, const Standard_Real theParameter,
-                                     Standard_Real & theResult, Standard_Integer & theErrorCode) const
-{
-    Handle(TColgp_HArray1OfPnt2d) points = new TColgp_HArray1OfPnt2d(1, old_parameters->Length());
-    for (int param_idx = 1; param_idx <= old_parameters->Length(); ++param_idx) {
-        points->SetValue(param_idx, gp_Pnt2d(old_parameters->Value(param_idx), 0));
-    }
-
-    Geom2dAPI_Interpolate interpolationObject(points, new_parameters, false, 1e-15);
-    interpolationObject.Perform();
-
-    // check that interpolation was successful
-    assert(interpolationObject.IsDone());
-
-    Handle(Geom2d_BSplineCurve) curve = interpolationObject.Curve();
-
-    gp_Pnt2d resulting_point = curve->Value(theParameter);
-    theResult = resulting_point.X();
-
-}
-
 Handle(Geom_BSplineCurve) CTiglBSplineAlgorithms::reparametrizeBSplineContinuouslyApprox(const Handle(Geom_BSplineCurve) spline,
                                                                                          const std::vector<double>& old_parameters,
                                                                                          const std::vector<double>& new_parameters,
@@ -938,154 +757,6 @@ Handle(Geom_BSplineSurface) CTiglBSplineAlgorithms::pointsToSurface(const TColgp
     return interpolatingSurf;
 }
 
-void checkCurveNetworkCompatibility(const std::vector<Handle(Geom_BSplineCurve) >& profiles,
-                                    const std::vector<Handle(Geom_BSplineCurve) >& guides,
-                                    const std::vector<double>& intersection_params_spline_u,
-                                    const std::vector<double>& intersection_params_spline_v,
-                                    double tol = 3e-4)
-{
-    // find out the 'average' scale of the B-splines in order to being able to handle a more approximate dataset and find its intersections
-    double splines_scale = 0.5 * (CTiglBSplineAlgorithms::scale(profiles)+ CTiglBSplineAlgorithms::scale(guides));
-
-    if (std::abs(intersection_params_spline_u.front()) > splines_scale * tol || std::abs(intersection_params_spline_u.back() - 1.) > splines_scale * tol) {
-        throw tigl::CTiglError("WARNING: B-splines in u-direction mustn't stick out, spline network must be 'closed'!");
-    }
-
-    if (std::abs(intersection_params_spline_v.front()) > splines_scale * tol || std::abs(intersection_params_spline_v.back() - 1.) > splines_scale * tol) {
-        throw tigl::CTiglError("WARNING: B-splines in v-direction mustn't stick out, spline network must be 'closed'!");
-    }
-
-    // check compatibilty of network
-    for (size_t u_param_idx = 0; u_param_idx < intersection_params_spline_u.size(); ++u_param_idx) {
-        double spline_u_param = intersection_params_spline_u[u_param_idx];
-        const Handle(Geom_BSplineCurve)& spline_v = guides[u_param_idx];
-        for (size_t v_param_idx = 0; v_param_idx < intersection_params_spline_v.size(); ++v_param_idx) {
-            const Handle(Geom_BSplineCurve)& spline_u = profiles[v_param_idx];
-            double spline_v_param = intersection_params_spline_v[v_param_idx];
-
-            gp_Pnt p_prof = spline_u->Value(spline_u_param);
-            gp_Pnt p_guid = spline_v->Value(spline_v_param);
-            double distance = p_prof.Distance(p_guid);
-
-            if (distance > splines_scale * tol) {
-                throw tigl::CTiglError("B-spline network is incompatible (e.g. wrong parametrization) or intersection parameters are in a wrong order!");
-            }
-        }
-    }
-}
-
-Handle(Geom_BSplineSurface) CTiglBSplineAlgorithms::createGordonSurface(const std::vector<Handle(Geom_BSplineCurve) >& profiles,
-                                                                        const std::vector<Handle(Geom_BSplineCurve) >& guides,
-                                                                        const Handle(TColStd_HArray1OfReal) intersection_params_spline_u,
-                                                                        const Handle(TColStd_HArray1OfReal) intersection_params_spline_v)
-{
-    // check whether there are any u-directional and v-directional B-splines in the vectors
-    if (profiles.size() < 2) {
-        throw CTiglError("There must be at least two profiles for the gordon surface.", TIGL_MATH_ERROR);
-    }
-
-    if (guides.size()  < 2) {
-        throw CTiglError("There must be at least two guides for the gordon surface.", TIGL_MATH_ERROR);
-    }
-
-    // check B-spline parametrization (should be from 0 to 1):
-    for (unsigned int profileIdx = 0; profileIdx < profiles.size(); ++profileIdx) {
-        assertRange(profiles[profileIdx], 0., 1., 1e-5);
-    }
-
-    for (unsigned int guideIdx = 0; guideIdx < guides.size(); ++guideIdx) {
-        assertRange(guides[guideIdx], 0., 1., 1e-5);
-    }
-
-    // TODO: Do we really need to check compatibility?
-    // We don't need to do this, if the curves were reparametrized before
-    // In this case, they might be even incompatible, as the curves have been approximated
-    checkCurveNetworkCompatibility(profiles, guides,
-                                   toVector(intersection_params_spline_u->Array1()),
-                                   toVector(intersection_params_spline_v->Array1()));
-
-    // setting everything up for creating Tensor Product Surface by interpolating intersection points of profiles and guides with B-Spline surface
-    // find the intersection points:
-    TColgp_Array2OfPnt intersection_pnts(1, intersection_params_spline_u->Upper(), 1, intersection_params_spline_v->Upper());
-
-    // use splines in u-direction to get intersection points
-    for (size_t spline_idx = 0; spline_idx < profiles.size(); ++spline_idx) {
-        for (int intersection_idx = intersection_params_spline_u->Lower(); intersection_idx <= intersection_params_spline_u->Upper(); ++intersection_idx) {
-            Handle(Geom_BSplineCurve) spline_u = profiles[spline_idx];
-            double parameter = intersection_params_spline_u->Value(intersection_idx);
-            intersection_pnts(intersection_idx, static_cast<Standard_Integer>(spline_idx + 1)) = spline_u->Value(parameter);
-        }
-    }
-
-    // check, whether to build a closed continous surface
-    double curve_u_tolerance = REL_TOL_CLOSED * scale(guides);
-    double curve_v_tolerance = REL_TOL_CLOSED * scale(profiles);
-    double tp_tolerance = REL_TOL_CLOSED * scale(intersection_pnts);
-
-    bool makeUClosed = isUDirClosed(intersection_pnts, tp_tolerance) && guides.front()->IsEqual(guides.back(), curve_u_tolerance);
-    bool makeVClosed = isVDirClosed(intersection_pnts, tp_tolerance) && profiles.front()->IsEqual(profiles.back(), curve_v_tolerance);
-
-    // Skinning in v-direction with u directional B-Splines
-    Handle(Geom_BSplineSurface) surface_v = curvesToSurface(profiles, toVector(intersection_params_spline_v->Array1()), makeVClosed);
-    // therefore reparametrization before this method
-
-    // Skinning in u-direction with v directional B-Splines
-    Handle(Geom_BSplineSurface) surface_u_unflipped = curvesToSurface(guides, toVector(intersection_params_spline_u->Array1()), makeUClosed);
-
-    // flipping of the surface in v-direction; flipping is redundant here, therefore the next line is a comment!
-    Handle(Geom_BSplineSurface) surface_u = flipSurface(surface_u_unflipped);
-
-    // if there are too little points for degree in u-direction = 3 and degree in v-direction=3 creating an interpolation B-spline surface isn't possible in Open CASCADE
-
-    // Open CASCADE doesn't have a B-spline surface interpolation method where one can give the u- and v-directional parameters as arguments
-    Handle(Geom_BSplineSurface) tensorProdSurf = CTiglBSplineAlgorithms::pointsToSurface(intersection_pnts,
-                                                                                         toVector(intersection_params_spline_u->Array1()),
-                                                                                         toVector(intersection_params_spline_v->Array1()),
-                                                                                         makeUClosed, makeVClosed);
-
-    // match degree of all three surfaces
-    Standard_Integer degreeU = std::max(std::max(surface_u->UDegree(),
-                                                 surface_v->UDegree()),
-                                                 tensorProdSurf->UDegree());
-
-    Standard_Integer degreeV = std::max(std::max(surface_u->VDegree(),
-                                                 surface_v->VDegree()),
-                                                 tensorProdSurf->VDegree());
-
-    // check whether degree elevation is necessary (does method elevate_degree_u()) and if yes, elevate degree
-    surface_u->IncreaseDegree(degreeU, degreeV);
-    surface_v->IncreaseDegree(degreeU, degreeV);
-    tensorProdSurf->IncreaseDegree(degreeU, degreeV);
-
-    std::vector<Handle(Geom_BSplineSurface)> surfaces_vector_unmod;
-    surfaces_vector_unmod.push_back(surface_u);
-    surfaces_vector_unmod.push_back(surface_v);
-    surfaces_vector_unmod.push_back(tensorProdSurf);
-
-    // create common knot vector for all three surfaces
-    std::vector<Handle(Geom_BSplineSurface)> surfaces_vector = createCommonKnotsVectorSurface(surfaces_vector_unmod);
-
-    assert(surfaces_vector.size() == 3);
-
-    surface_u = surfaces_vector[0];
-    surface_v = surfaces_vector[1];
-    tensorProdSurf = surfaces_vector[2];
-
-    // creating the Gordon Surface = s_u + s_v - tps
-    // We reuse the u surface for storing the result
-
-    for (int cp_u_idx = 1; cp_u_idx <= surface_u->NbUPoles(); ++cp_u_idx) {
-        for (int cp_v_idx = 1; cp_v_idx <= surface_u->NbVPoles(); ++cp_v_idx) {
-            gp_Pnt cp_surf_u = surface_u->Pole(cp_u_idx, cp_v_idx);
-            gp_Pnt cp_surf_v = surface_v->Pole(cp_u_idx, cp_v_idx);
-            gp_Pnt cp_tensor = tensorProdSurf->Pole(cp_u_idx, cp_v_idx);
-
-            surface_u->SetPole(cp_u_idx, cp_v_idx, cp_surf_u.XYZ() + cp_surf_v.XYZ() - cp_tensor.XYZ());
-        }
-    }
-
-    return surface_u;
-}
 
 std::vector<std::pair<double, double> > CTiglBSplineAlgorithms::intersections(const Handle(Geom_BSplineCurve) spline1, const Handle(Geom_BSplineCurve) spline2, double tolerance) {
     // light weight simple minimizer
@@ -1162,47 +833,6 @@ double CTiglBSplineAlgorithms::scale(const Handle(Geom_BSplineCurve)& spline)
     return scale;
 }
 
-void CTiglBSplineAlgorithms::eliminateInaccuraciesNetworkIntersections(const std::vector<Handle(Geom_BSplineCurve)> & sorted_splines_u, const std::vector<Handle(Geom_BSplineCurve)> & sorted_splines_v, math_Matrix & intersection_params_u, math_Matrix & intersection_params_v) {
-
-    // eliminate small inaccuracies of the intersection parameters:
-
-    // first intersection
-    for (unsigned int spline_u_idx = 0; spline_u_idx < sorted_splines_u.size(); ++spline_u_idx) {
-        if (std::abs(intersection_params_u(spline_u_idx, 0) - sorted_splines_u[0]->Knot(1)) < 0.001) {
-            if (std::abs(sorted_splines_u[0]->Knot(1)) < 1e-10) {
-                intersection_params_u(spline_u_idx, 0) = 0;
-            }
-            else {
-                intersection_params_u(spline_u_idx, 0) = sorted_splines_u[0]->Knot(1);
-            }
-        }
-    }
-
-    for (unsigned int spline_v_idx = 0; spline_v_idx < sorted_splines_v.size(); ++spline_v_idx) {
-        if (std::abs(intersection_params_v(0, spline_v_idx) - sorted_splines_v[0]->Knot(1)) < 0.001) {
-            if (std::abs(sorted_splines_v[0]->Knot(1)) < 1e-10) {
-                intersection_params_v(0, spline_v_idx) = 0;
-            }
-            else {
-                intersection_params_v(0, spline_v_idx) = sorted_splines_v[0]->Knot(1);
-            }
-        }
-    }
-
-    // last intersection
-    for (unsigned int spline_u_idx = 0; spline_u_idx < sorted_splines_u.size(); ++spline_u_idx) {
-        if (std::abs(intersection_params_u(spline_u_idx, sorted_splines_v.size() - 1) - sorted_splines_u[0]->Knot(sorted_splines_u[0]->NbKnots())) < 0.001) {
-            intersection_params_u(spline_u_idx, sorted_splines_v.size() - 1) = sorted_splines_u[0]->Knot(sorted_splines_u[0]->NbKnots());
-        }
-    }
-
-    for (unsigned int spline_v_idx = 0; spline_v_idx < sorted_splines_v.size(); ++spline_v_idx) {
-        if (std::abs(intersection_params_v(sorted_splines_u.size() - 1, spline_v_idx) - sorted_splines_v[0]->Knot(sorted_splines_v[0]->NbKnots())) < 0.001) {
-            intersection_params_v(sorted_splines_u.size() - 1, spline_v_idx) = sorted_splines_v[0]->Knot(sorted_splines_v[0]->NbKnots());
-        }
-    }
-}
-
 void CTiglBSplineAlgorithms::reparametrizeBSpline(Geom_BSplineCurve& spline, double umin, double umax, double tol)
 {
     if (std::abs(spline.Knot(1) - umin) > tol || std::abs(spline.Knot(spline.NbKnots()) - umax) > tol) {
@@ -1211,242 +841,6 @@ void CTiglBSplineAlgorithms::reparametrizeBSpline(Geom_BSplineCurve& spline, dou
         BSplCLib::Reparametrize (0., 1., aKnots);
         spline.SetKnots (aKnots);
     }
-}
-
-void computeIntersections(const std::vector<Handle(Geom_BSplineCurve)>& profiles, const std::vector<Handle(Geom_BSplineCurve)>& guides, math_Matrix& intersection_params_u, math_Matrix& intersection_params_v)
-{
-    for (unsigned int spline_u_idx = 0; spline_u_idx < profiles.size(); ++spline_u_idx) {
-        for (unsigned int spline_v_idx = 0; spline_v_idx < guides.size(); ++spline_v_idx) {
-            std::vector<std::pair<double, double> > intersection_params_vector = CTiglBSplineAlgorithms::intersections(profiles[spline_u_idx], guides[spline_v_idx]);
-
-            if (intersection_params_vector.size() < 1) {
-                throw tigl::CTiglError("U-directional B-spline and v-directional B-spline don't intersect each other!");
-            }
-
-            else if (intersection_params_vector.size() == 1) {
-                intersection_params_u(spline_u_idx, spline_v_idx) = intersection_params_vector[0].first;
-                intersection_params_v(spline_u_idx, spline_v_idx) = intersection_params_vector[0].second;
-            }
-                // for closed curves
-            else if (intersection_params_vector.size() == 2) {
-
-                // only the u-directional B-spline curves are closed
-                if (profiles[0]->IsClosed()) {
-
-                    if (spline_v_idx == 0) {
-                        intersection_params_u(spline_u_idx, spline_v_idx) = std::min(intersection_params_vector[0].first, intersection_params_vector[1].first);
-                    }
-                    else if (spline_v_idx == guides.size() - 1) {
-                        intersection_params_u(spline_u_idx, spline_v_idx) = std::max(intersection_params_vector[0].first, intersection_params_vector[1].first);
-                    }
-
-                    // intersection_params_vector[0].second == intersection_params_vector[1].second
-                    intersection_params_v(spline_u_idx, spline_v_idx) = intersection_params_vector[0].second;
-                }
-
-                // only the v-directional B-spline curves are closed
-                if (guides[0]->IsClosed()) {
-
-                    if (spline_u_idx == 0) {
-                        intersection_params_v(spline_u_idx, spline_v_idx) = std::min(intersection_params_vector[0].second, intersection_params_vector[1].second);
-                    }
-                    else if (spline_u_idx == profiles.size() - 1) {
-                        intersection_params_v(spline_u_idx, spline_v_idx) = std::max(intersection_params_vector[0].second, intersection_params_vector[1].second);
-                    }
-                    // intersection_params_vector[0].first == intersection_params_vector[1].first
-                    intersection_params_u(spline_u_idx, spline_v_idx) = intersection_params_vector[0].first;
-                }
-
-//                // TODO: both u-directional splines and v-directional splines are closed
-//               else if (intersection_params_vector.size() == 4) {
-
-//                }
-            }
-
-            else if (intersection_params_vector.size() > 2) {
-                throw tigl::CTiglError("U-directional B-spline and v-directional B-spline have more than two intersections with each other!");
-            }
-        }
-    }    
-}
-
-Handle(Geom_BSplineSurface) CTiglBSplineAlgorithms::createGordonSurfaceGeneral(const std::vector<Handle(Geom_BSplineCurve) >& splines_u_vector_in,
-                                                                               const std::vector<Handle(Geom_BSplineCurve) >& splines_v_vector_in)
-{
-
-    std::vector<Handle(Geom_BSplineCurve) > profiles(splines_u_vector_in);
-    std::vector<Handle(Geom_BSplineCurve) > guides(splines_v_vector_in);
-
-    // check whether there are any u-directional and v-directional B-splines in the vectors
-    if (profiles.size() == 0) {
-        throw tigl::CTiglError("There are no profile curves!");
-    }
-
-    if (guides.size() == 0) {
-        throw tigl::CTiglError("There are no guide curves!");
-    }
-
-    // reparametrize into [0,1]
-    for (unsigned int spline_u_idx = 0; spline_u_idx < profiles.size(); ++spline_u_idx) {
-        reparametrizeBSpline(*profiles[spline_u_idx], 0., 1., 1e-15);
-    }
-
-    for (unsigned int spline_v_idx = 0; spline_v_idx < guides.size(); ++spline_v_idx) {
-        reparametrizeBSpline(*guides[spline_v_idx], 0., 1., 1e-15);
-    }
-    // now the parameter range of all  profiles and guides is [0, 1]
-
-    // now find all intersections of all B-splines with each other
-    math_Matrix intersection_params_u(0, profiles.size() - 1, 0, guides.size() - 1);
-    math_Matrix intersection_params_v(0, profiles.size() - 1, 0, guides.size() - 1);
-    computeIntersections(profiles, guides, intersection_params_u, intersection_params_v);
-
-    // sort intersection_params_u and intersection_params_v and u-directional and v-directional B-spline curves
-    tigl::CTiglCurveNetworkSorter sorterObj(std::vector<Handle(Geom_Curve)>(profiles.begin(), profiles.end()),
-                                            std::vector<Handle(Geom_Curve)>(guides.begin(), guides.end()),
-                                            intersection_params_u,
-                                            intersection_params_v);
-    sorterObj.Perform();
-
-    // get the sorted matrices and vectors
-    intersection_params_u = sorterObj.ProfileIntersectionParms();
-    intersection_params_v = sorterObj.GuideIntersectionParms();
-
-    // copy sorted curves back into out curves array
-    std::transform(sorterObj.Profiles().begin(), sorterObj.Profiles().end(), profiles.begin(), Handle(Geom_BSplineCurve)::DownCast);
-    std::transform(sorterObj.Guides().begin(), sorterObj.Guides().end(), guides.begin(), Handle(Geom_BSplineCurve)::DownCast);
-
-    // eliminate small inaccuracies of the intersection parameters:
-    CTiglBSplineAlgorithms::eliminateInaccuraciesNetworkIntersections(profiles, guides, intersection_params_u, intersection_params_v);
-
-    TColStd_Array1OfReal average_intersection_params_u(1, guides.size());
-    for (unsigned int spline_v_idx = 1; spline_v_idx <= guides.size(); ++spline_v_idx) {
-        double sum = 0;
-        for (unsigned int spline_u_idx = 1; spline_u_idx <= profiles.size(); ++spline_u_idx) {
-            sum += intersection_params_u(spline_u_idx - 1, spline_v_idx - 1);
-        }
-        average_intersection_params_u(spline_v_idx) = sum / profiles.size();
-    }
-
-    TColStd_Array1OfReal average_intersection_params_v(1, profiles.size());
-    for (unsigned int spline_u_idx = 1; spline_u_idx <= profiles.size(); ++spline_u_idx) {
-        double sum = 0;
-        for (unsigned int spline_v_idx = 1; spline_v_idx <= guides.size(); ++spline_v_idx) {
-            sum += intersection_params_v(spline_u_idx - 1, spline_v_idx - 1); // TODO: scientific approach for power
-        }
-        average_intersection_params_v(spline_u_idx) = sum / guides.size();
-    }
-
-
-    if (average_intersection_params_u(1) > 1e-5 || average_intersection_params_v(1) > 1e-5) {
-        throw CTiglError("At least one B-splines has no intersection at the beginning.");
-    }
-
-    // Get maximum number of control points to figure out detail of spline
-    int max_cp_u = 0, max_cp_v = 0;
-    for(std::vector<Handle(Geom_BSplineCurve)>::const_iterator it = profiles.begin(); it != profiles.end(); ++it) {
-        max_cp_u = std::max(max_cp_u, (*it)->NbPoles());
-    }
-    for(std::vector<Handle(Geom_BSplineCurve)>::const_iterator it = guides.begin(); it != guides.end(); ++it) {
-        max_cp_v = std::max(max_cp_v, (*it)->NbPoles());
-    }
-
-    // we want to use at least 10 and max 80 control points to be able to reparametrize the geometry properly
-    size_t mincp = 10;
-    size_t maxcp = 80;
-
-    // since we interpolate the intersections, we cannot use fewer control points than curves
-    // We need to add two since we want c2 continuity, which adds two equations
-    size_t min_u = std::max(guides.size() + 2, mincp);
-    size_t min_v = std::max(profiles.size() + 2, mincp);
-
-    size_t max_u = std::max(min_u, maxcp);
-    size_t max_v = std::max(min_v, maxcp);
-
-    max_cp_u = Clamp(max_cp_u + 10, static_cast<int>(min_u), static_cast<int>(max_u));
-    max_cp_v = Clamp(max_cp_v + 10, static_cast<int>(min_v), static_cast<int>(max_v));
-
-    // reparametrize u-directional B-splines
-    std::vector<Handle(Geom_BSplineCurve)> reparam_splines_u;
-    for (unsigned int spline_u_idx = 0; spline_u_idx < profiles.size(); ++spline_u_idx) {
-
-        TColStd_Array1OfReal old_parameters(1, guides.size());
-        for (unsigned int spline_v_idx = 1; spline_v_idx <= guides.size(); ++spline_v_idx) {
-            old_parameters(spline_v_idx) = intersection_params_u(spline_u_idx, spline_v_idx - 1);
-        }
-
-        // eliminate small inaccuracies at the first knot
-        if (std::abs(old_parameters(1)) < 1e-5) {
-            old_parameters(1) = 0;
-        }
-
-        if (std::abs(average_intersection_params_u(1)) < 1e-5) {
-            average_intersection_params_u(1) = 0;
-        }
-
-        // eliminate small inaccuracies at the last knot
-        if (std::abs(old_parameters(old_parameters.Length()) - 1) < 1e-5) {
-            old_parameters(old_parameters.Length()) = 1;
-        }
-
-        if (std::abs(average_intersection_params_u(average_intersection_params_u.Length()) - 1) < 1e-5) {
-            average_intersection_params_u(average_intersection_params_u.Length()) = 1;
-        }
-
-
-        Handle(Geom_BSplineCurve) reparam_spline_u = CTiglBSplineAlgorithms::reparametrizeBSplineContinuouslyApprox(profiles[spline_u_idx], toVector(old_parameters), toVector(average_intersection_params_u), max_cp_u);
-        reparam_splines_u.push_back(reparam_spline_u);
-    }
-
-    // reparametrize v-directional B-splines
-    std::vector<Handle(Geom_BSplineCurve)> reparam_splines_v;
-    for (unsigned int spline_v_idx = 0; spline_v_idx < guides.size(); ++spline_v_idx) {
-
-        TColStd_Array1OfReal old_parameters(1, profiles.size());
-        for (unsigned int spline_u_idx = 1; spline_u_idx <= profiles.size(); ++spline_u_idx) {
-            old_parameters(spline_u_idx) = intersection_params_v(spline_u_idx - 1, spline_v_idx);
-        }
-
-        // eliminate small inaccuracies at the first knot
-        if (std::abs(old_parameters(1)) < 1e-5) {
-            old_parameters(1) = 0;
-        }
-
-        if (std::abs(average_intersection_params_v(1)) < 1e-5) {
-            average_intersection_params_v(1) = 0;
-        }
-
-        // eliminate small inaccuracies at the last knot
-        if (std::abs(old_parameters(old_parameters.Length()) - 1) < 1e-5) {
-            old_parameters(old_parameters.Length()) = 1;
-        }
-
-        if (std::abs(average_intersection_params_v(average_intersection_params_v.Length()) - 1) < 1e-5) {
-            average_intersection_params_v(average_intersection_params_v.Length()) = 1;
-        }
-
-        Handle(Geom_BSplineCurve) reparam_spline_v = CTiglBSplineAlgorithms::reparametrizeBSplineContinuouslyApprox(guides[spline_v_idx],
-                                                                                                                    toVector(old_parameters),
-                                                                                                                    toVector(average_intersection_params_v),
-                                                                                                                    max_cp_v);
-        reparam_splines_v.push_back(reparam_spline_v);
-    }
-
-    // splines are reparametrized, but don't have common knot vector yet, in this sense not compatible
-
-    // convert types of average_intersection_params
-    Handle(TColStd_HArray1OfReal) av_intersection_params_u = new TColStd_HArray1OfReal(1, average_intersection_params_u.Length());
-    for (int i = 1; i <= average_intersection_params_u.Length(); ++i) {
-        av_intersection_params_u->SetValue(i, average_intersection_params_u(i));
-    }
-
-    Handle(TColStd_HArray1OfReal) av_intersection_params_v = new TColStd_HArray1OfReal(1, average_intersection_params_v.Length());
-    for (int i = 1; i <= average_intersection_params_v.Length(); ++i) {
-        av_intersection_params_v->SetValue(i, average_intersection_params_v(i));
-    }
-
-    Handle(Geom_BSplineSurface) gordonSurface = CTiglBSplineAlgorithms::createGordonSurface(reparam_splines_u, reparam_splines_v, av_intersection_params_u, av_intersection_params_v);
-    return gordonSurface;
 }
 
 math_Matrix CTiglBSplineAlgorithms::bsplineBasisMat(int degree, const TColStd_Array1OfReal& knots, const TColStd_Array1OfReal& params, unsigned int derivOrder)
